backend/requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
python-dotenv
httpx
pyjwt
python-multipart
